requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic-settings==2.1.0
pydantic[email]>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==8.0.0
pytest-asyncio==0.23.0
prometheus-fastapi-instrumentator==7.0.0
httpx==0.26.0
aiosqlite==0.19.0
email-validator>=2.1.0
ruff>=0.2.0
pip-audit>=2.7.0
pytest-cov>=4.0.0